Add files via upload
Most standard Pomodoro timers are fixed on a 25 minute on/5 minute off schedule. This restricts personal preferences.

You can use a Google Stopwatch, but it is laborious to keep restarting new timers.

The purpose of this programme is to set up a series of Pomodoro timers that suits your personal preferences without needing to start
each individual timer.

Each timer does not display the amount of time left on itself. This design choice was made so that you cannot distract yourself by
watching a clock tick away.